tota-benchmarking/pyproject.toml
[tool.poetry]
name = "tota-benchmarking"
version = "0.0.1"
description = ""
authors = ["Sarvam <ritvik@sarvam.ai>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pydantic = "^2.10.6"
datasets = "^3.3.2"
httpx = "0.27.2"
openai = "^1.66.2"
cachetools = "^5.5.2"
huggingface-hub = "^0.24.0"
whatsapp-widgets = {version = "^0.3.68", source = "sarvam-python-ci"}
sarvam-stream = {version = "^0.3.9", source = "sarvam-python-ci"}
sarvam-agents-sdk = {version = "^0.2.144", source = "sarvam-python-ci"}
sarvam-datatypes = {version = "^0.4.20", source = "sarvam-python-ci"}
sarvam-app-stream = {version = "^0.5.102", source = "sarvam-python-ci"}
sarvam-tool = {version = "^0.2.13", source = "sarvam-python-ci"}
sarvam-observability = {version = "^0.5.0", source = "sarvam-python-ci"}
sarvam-llm = {version = "^0.5.19", source = "sarvam-python-ci"}
sarvam-kb = {version = "^0.1.13", source = "sarvam-python-ci"}
sarvam-agent = {version = "^0.4.10", source = "sarvam-python-ci"}
anthropic = {extras = ["vertex"], version = "^0.52.0"}
google-genai = "^1.0.0"
groq = "^0.31.0"
streamlit = "^1.45.0"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
flake8 = "^6.1.0"
mypy = "^1.5.1"
pytest = "^7.4.2"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"
autodoc-pydantic = "^2.0.1"
pytest-asyncio = "^0.23.7"
types-tqdm = "^4.67.0.20250319"
ipykernel = "^6.29.5"

[[tool.poetry.source]]
name = "sarvam-python-ci"
url = "https://europe-west4-python.pkg.dev/gpu-reservation-sarvam/sarvam-python-ci/simple/"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
